+++ src/components/POLine/PaymentTerms/PaymentTermsView/PaymentTermsView.js
@@ -1,0 +1,3 @@
+export const PaymentTermsView = () => {
+  return 'PaymentTermsView';
+};