+++ src/components/POLine/PaymentTerms/index.js
@@ -1,0 +1,5 @@
+export {
+  PaymentTermsForm,
+  PaymentTermsFormContainer,
+} from './PaymentTermsForm';
+export { PaymentTermsView } from './PaymentTermsView';