+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fix the `budgetExpenseClassNotFound` error handling. Refs UIOR-1552.
 * *BREAKING* Update CQL queries to use the new indices. Refs UIOR-1519.
 * Update the "Ongoing order information" for PO Line to define multi-year payments. Refs UIOR-1528.
+* Implement "Payment terms" form. Refs UIOR-1530.
 
 ## [9.0.4](https://github.com/folio-org/ui-orders/tree/v9.0.4) (2026-05-26)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v9.0.3...v9.0.4)